requirements.txt
fastapi
uvicorn
jinja2

python-dotenv

#infrastructure
azure-mgmt-resource
azure-identity
